+++ src/fork_exec.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "Error.h"
 
-int fork_exec(char* file, char* argv[], Error* error);
+int fork_exec(char* file, char* argv[], int* pid, Error* error);