+++ test/unknown_path.t
@@ -6,3 +6,4 @@ Complain about non-existent path:
 
   $ show doesnt_exist
   ls: cannot access doesnt_exist: No such file or directory
+  [1]